Ver 2.0.4
Bug Fixes:
Two commits took place.

The first commit was targeted to squash a bug where on mobile view the width of the landing page was 100 viewable height, rather than width

Second commit was targeted to fix website catalogue not presenting as expected on medium to large sized devices. issue is persistent on smaller 'mini' devices. This patch will be next.